% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2jags.R
\name{csv2jags}
\alias{csv2jags}
\title{convert CSV in a specific format to a jagsdata file}
\usage{
csv2jags(type, package = "barca", dodump = TRUE)
}
\arguments{
\item{type}{character string in \code{c("TypeIII", "TypeIa", "TypeV")}}

\item{package}{name of package in which the file will be found under \code{csv/} in installed folder}

\item{dodump}{logical indicating that the base::dump function will be used to create a .jagsdata file, with prefix given by the value of \code{type}}
}
\description{
convert CSV in a specific format to a jagsdata file
}
