% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editBugPars.R
\name{editBugPars}
\alias{editBugPars}
\title{convenience function for setting hyperparameters and point density}
\usage{
editBugPars(blines, beta_a = 25, beta_b = 2500, ini_x = 0.1,
  del_x = 0.1, npts = 20)
}
\arguments{
\item{blines}{character vector of bugs text}

\item{beta_a}{numeric beta hyperparameter value}

\item{beta_b}{numeric beta hyperparameter value}

\item{ini_x}{initial value of sequence of concentrations for pointwise posterior generation}

\item{del_x}{gap value for sequence of concentrations for pointwise posterior generation}

\item{npts}{number of concentration values to survey}
}
\description{
convenience function for setting hyperparameters and point density
}
